Add splitM/N feature to the F16 kernel
Kernel wrapper has an option to split the matmul in M/N dimension
as preset size. And sub-tile matmul is calculated using corresponding
partition of W/X tensor using .view and also written to the view on Y.
So it shouldn't involve any data copy or accumulation like split-K.
This helps while persistent kernel is not supported and could give us
better idea for cache efficient way of sub-tile walk-through.
Enabled with TRITON_HIP_MNSPLIT env_var, '1' is more aggressive preset
and '2' excludes some shapes which might not be beneficial.